Updates for HW3
