+++ src/main/java/driverutils/ChromeDriverManager.java
@@ -9,19 +9,14 @@ import org.testng.annotations.BeforeClass;
 
 public class ChromeDriverManager implements Driverinterface {
 
-    public static WebDriver driver;
-
-    @BeforeClass
     public WebDriver setupChromeDriver() {
+        WebDriver driver = null;
         if (driver == null) {
             WebDriverManager.chromedriver().setup();
             //1 Open BR
             driver = new ChromeDriver();
             driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
             driver.manage().window().maximize();
-
-            //2 Navigate
-            driver.navigate().to("https://jdi-testing.github.io/jdi-light/index.html");
         }
         return driver;
     }
