+++ src/main/resources/HW3ExerciseTestSuite.xml
@@ -5,7 +5,9 @@
 
     <test name = "Exercises Hw3">
         <classes>
-            <class name = "hw.three.ExercisesHw3" />
+            <class name = "hw.three.ExerciseHw3ex1" />
+            <class name = "hw.three.ExerciseHw3ex2" />
+            <class name = "hw.three.ExerciseHw3exFluent" />
         </classes>
     </test>
