+++ src/test/java/hw/three/ExerciseHw3ex1.java
@@ -1,0 +1,24 @@
+package hw.three;
+
+import static org.testng.Assert.assertEquals;
+
+import org.testng.annotations.Test;
+import pages.IndexPage;
+
+
+public class ExerciseHw3ex1 extends AbstractSettingClass {
+
+    IndexPage indexPage;
+
+    @Test
+    public void ex1Test() {
+        indexPage = new IndexPage(driver);
+        indexPage.openIndexPage();
+        assertEquals(driver.getTitle(),
+                "Home Page");
+        indexPage.getHeaderMenu().checkAllHeaderItems();
+        indexPage.checkPageIconsAndText();
+        indexPage.iframeButtonCheck();
+        indexPage.getLeftSideMenu().checkAllLeftMenuItems();
+    }
+}
